Split GUI drag and drop actions
